backend/requirement.txt
fastapi
python-jose
mysql-connector
passlib
bcrypt